requirements.txt
accelerate==0.21.0
datasets==2.15.0
diffusers==0.25.0
einops==0.8.0
glfw==2.7.0
gradio==4.44.0
huggingface_hub==0.19.4
imageio==2.33.0
kiui==0.2.13
matplotlib==3.7.2
numpy==2.1.1
omegaconf==2.3.0
open3d==0.14.1
open_clip_torch==2.24.0
opencv_contrib_python==4.8.1.78
opencv_python==4.8.0.74
opencv_python_headless==4.8.1.78
Pillow==10.4.0
pygltflib==1.16.2
rembg==2.0.53
scipy==1.14.1
setuptools==68.0.0
tqdm==4.65.0
transformers==4.37.0.dev0
trimesh==3.23.1
xatlas==0.0.7
git+https://github.com/NVlabs/nvdiffrast.git
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
git+https://github.com/facebookresearch/xformers.git@main#egg=xformers